+++ curseforge-client/src/models/fingerprint_matches_result.rs
@@ -17,21 +17,21 @@ pub struct FingerprintMatchesResult {
     #[serde(rename = "exactMatches")]
     pub exact_matches: Vec<models::FingerprintMatch>,
     #[serde(rename = "exactFingerprints")]
-    pub exact_fingerprints: Vec<usize>,
+    pub exact_fingerprints: Vec<u64>,
     #[serde(rename = "partialMatches")]
     pub partial_matches: Vec<models::FingerprintMatch>,
     #[serde(rename = "partialMatchFingerprints")]
     pub partial_match_fingerprints: serde_json::Value,
     #[serde(rename = "additionalProperties")]
-    pub additional_properties: Vec<usize>,
+    pub additional_properties: Vec<u64>,
     #[serde(rename = "installedFingerprints")]
-    pub installed_fingerprints: Vec<usize>,
+    pub installed_fingerprints: Vec<u64>,
     #[serde(rename = "unmatchedFingerprints")]
-    pub unmatched_fingerprints: Vec<usize>,
+    pub unmatched_fingerprints: Vec<u64>,
 }
 
 impl FingerprintMatchesResult {
-    pub fn new(is_cache_built: bool, exact_matches: Vec<models::FingerprintMatch>, exact_fingerprints: Vec<usize>, partial_matches: Vec<models::FingerprintMatch>, partial_match_fingerprints: serde_json::Value, additional_properties: Vec<usize>, installed_fingerprints: Vec<usize>, unmatched_fingerprints: Vec<usize>) -> Self {
+    pub fn new(is_cache_built: bool, exact_matches: Vec<models::FingerprintMatch>, exact_fingerprints: Vec<u64>, partial_matches: Vec<models::FingerprintMatch>, partial_match_fingerprints: serde_json::Value, additional_properties: Vec<u64>, installed_fingerprints: Vec<u64>, unmatched_fingerprints: Vec<u64>) -> Self {
         Self {
             is_cache_built,
             exact_matches,
